% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cb_games}
\alias{cb_games}
\title{Calvinball League Game Results}
\format{
A tibble with 150 rows and 9 variables:
\describe{
  \item{season}{Season number (1-3)}
  \item{game_id}{Unique game identifier}
  \item{date}{Date of the game}
  \item{team_home}{Home team ID}
  \item{team_away}{Away team ID}
  \item{score_home}{Home team score (can be negative!)}
  \item{score_away}{Away team score (can be negative!)}
  \item{scoring_type}{The scoring system used for this game (changes each game)}
  \item{winner}{Team ID of the winner, or NA for ties}
}
}
\source{
Synthetically generated data inspired by Calvin and Hobbes
}
\usage{
cb_games
}
\description{
A dataset containing game results across 3 seasons of Calvinball.
Scores are intentionally chaotic, ranging from negative values to thousands,
reflecting the unpredictable nature of Calvinball.
}
\details{
Scoring types include: wickets, vortex_points, opposite_points,
time_zones, song_scores, invisible_sectors, spontaneous_declarations,
and base_nonsensicals.
}
\examples{
head(cb_games)

# Find games with negative scores
cb_games[cb_games$score_home < 0 | cb_games$score_away < 0, ]

}
\keyword{datasets}
