% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cb_player_summary}
\alias{cb_player_summary}
\title{Calvinball Player Career Summary}
\format{
A tibble with 30 rows and 9 variables:
\describe{
  \item{player_id}{Player identifier}
  \item{games_played}{Total games played}
  \item{avg_wickets}{Average wickets scored per game}
  \item{total_opposite_touchdowns}{Career total opposite touchdowns}
  \item{avg_style_points}{Average style points per game}
  \item{total_rule_declarations}{Career total spontaneous rule declarations}
  \item{avg_time_played}{Average minutes played per game}
  \item{player_name}{Player name}
  \item{team_id}{Current team ID}
}
}
\source{
Synthetically generated data inspired by Calvin and Hobbes
}
\usage{
cb_player_summary
}
\description{
Aggregated career statistics for each player.
}
\examples{
head(cb_player_summary)

# Top 5 players by average style points
head(cb_player_summary[order(-cb_player_summary$avg_style_points), ], 5)

}
\keyword{datasets}
