% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cb_players}
\alias{cb_players}
\title{Calvinball League Players}
\format{
A tibble with 30 rows and 3 variables:
\describe{
  \item{player_id}{Unique player identifier}
  \item{player_name}{Player name (creative Calvinball-style names)}
  \item{team_id}{Team identifier (1-6)}
}
}
\source{
Synthetically generated data inspired by Calvin and Hobbes
}
\usage{
cb_players
}
\description{
A dataset containing player information for the Calvinball League.
}
\examples{
head(cb_players)

}
\keyword{datasets}
