% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cb_team_records}
\alias{cb_team_records}
\title{Calvinball Team Records by Season}
\format{
A tibble with 18 rows and 6 variables:
\describe{
  \item{team_id}{Team identifier}
  \item{season}{Season number}
  \item{wins}{Number of wins}
  \item{losses}{Number of losses}
  \item{ties}{Number of ties}
  \item{team_name}{Team name}
}
}
\source{
Synthetically generated data inspired by Calvin and Hobbes
}
\usage{
cb_team_records
}
\description{
Win-loss-tie records for each team in each season.
}
\examples{
cb_team_records

}
\keyword{datasets}
