% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cb_teams}
\alias{cb_teams}
\title{Calvinball League Teams}
\format{
A tibble with 6 rows and 2 variables:
\describe{
  \item{team_id}{Unique team identifier}
  \item{team_name}{Team name (e.g., "Transmogrifiers", "Time Travelers")}
}
}
\source{
Synthetically generated data inspired by Calvin and Hobbes
}
\usage{
cb_teams
}
\description{
A dataset containing team information for the Calvinball League.
}
\examples{
cb_teams

}
\keyword{datasets}
